+++ src/index.ts
@@ -1,4 +1,4 @@
-export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse} from './rest/responseType';
+export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse, BlobAndFilenameResponse} from './rest/responseType';
 export {Cache} from './rest/cache';
 export {Body,BaseUrl,DefaultHeaders,DELETE,GET,HEAD,Header,Headers,JsonContentType,ParameterTransform,Path,POST,Produces,PUT,Query,QueryObject,ResponseTransform,RESTClient,FullHttpResponse,ReportProgress} from './rest/common';
 export {RestTransferStateModule} from './modules/restTransferState.module';